+++ CHANGELOG.md
@@ -32,6 +32,7 @@ We use *breaking* word for marking changes that are not backward compatible (rel
 ### Added
 
 - [#2671](https://github.com/thanos-io/thanos/pull/2671) Tools: bucket replicate now allows passing repeated `--compaction` and `--resolution` flags.
+- [#2754](https://github.com/thanos-io/thanos/pull/2671) UI: add stores page in the React UI.
 
 ## [v0.13.0](https://github.com/thanos-io/thanos/releases) - IN PROGRESS